requirements.txt
SQLAlchemy~=2.0.8
pydantic~=1.10.7
aiogram~=3.0.0b7
python-dotenv~=1.0.0
aiohttp~=3.8.4
lxml~=4.9.2
asyncpg~=0.27.0
aiodns~=3.0.0